Hide closure:run command (#280)
